# contracts/care-coordination-platform-openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Patient Engagement Platform API
  version: 1.0.0
  description: API-first contract examples for the care coordination platform.
servers:
  - url: https://api.healthcare.example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AppointmentRequest:
      type: object
      required: [patientId, providerId, scheduledAt, channel]
      properties:
        patientId: { type: string }
        providerId: { type: string }
        scheduledAt: { type: string, format: date-time }
        channel: { type: string }
      example: { patientId: "pat-1001", providerId: "prov-44", scheduledAt: "2026-06-01T09:30:00Z", channel: "VIDEO" }
    TeleconsultationCompleteRequest:
      type: object
      required: [consultationNotes]
      properties:
        consultationNotes: { type: string }
        followUpRequired: { type: boolean }
        nextFollowUpDate: { type: string, format: date-time }
      example:
        consultationNotes: Consultation completed; patient advised to continue current treatment and follow-up in 2 weeks.
        followUpRequired: true
        nextFollowUpDate: 2026-07-15T11:00:00Z
...
    AlertRequest:
      type: object
      required: [patientId, severity, triggerType, summary]
      properties:
        patientId: { type: string }
        deviceId:
          type: string
          description: Optional originating device identifier when alert is telemetry-driven
        severity: { type: string }
        triggerType: { type: string }
        metricValue:
          type: string
          description: Numeric telemetry value used for threshold evaluation when provided
        summary: { type: string }
      example: { patientId: "pat-1001", deviceId: "dev-22", severity: "HIGH", triggerType: "HEART_RATE", metricValue: "124", summary: "Heart rate exceeded configured threshold." }
...
    AppointmentResponse:
      type: object
      required: [id, status, patientId, providerId, scheduledAt, channel]
      properties:
        id: { type: string }
        status: { type: string }
        patientId: { type: string }
        providerId: { type: string }
        scheduledAt: { type: string }
        channel: { type: string }
      example: { id: "apt-2001", status: "BOOKED", patientId: "pat-1001", providerId: "prov-44", scheduledAt: "2026-06-01T09:30:00Z", channel: "VIDEO" }
    TeleconsultationResponse:
      type: object
      required: [sessionId, appointmentId, status, doctorJoinUrl, patientJoinUrl, startedAt, interactionLogs]
      properties:
        sessionId: { type: string }
        appointmentId: { type: string }
        status: { type: string }
        doctorJoinUrl: { type: string }
        patientJoinUrl: { type: string }
        startedAt: { type: string, format: date-time }
        joinedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        consultationNotes: { type: string }
        followUpRequired: { type: boolean }
        nextFollowUpDate: { type: string, format: date-time }
        interactionLogs:
          type: array
          items: { type: string }
      example:
        sessionId: tcs-5001
        appointmentId: apt-2001
        status: IN_PROGRESS
        doctorJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=DOCTOR
        patientJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=PATIENT
        startedAt: 2026-06-20T10:25:00Z
        joinedAt: 2026-06-20T10:30:30Z
        completedAt:
        consultationNotes:
        followUpRequired: false
        nextFollowUpDate:
        interactionLogs:
          - 2026-06-20T10:25:00Z | Teleconsultation initiated by doctor
          - 2026-06-20T10:30:30Z | Patient joined teleconsultation session
    AvailableSlotResponse:
      type: object
      required: [providerId, date, availableSlots]
      properties:
        providerId: { type: string }
        date: { type: string, format: date }
        availableSlots:
          type: array
          items: { type: string, format: date-time }
      example:
        providerId: prov-44
        date: 2026-06-01
        availableSlots:
          - 2026-06-01T09:00:00Z
          - 2026-06-01T09:30:00Z
          - 2026-06-01T10:00:00Z
...
    AlertResponse:
      type: object
      required: [id, status, patientId, severity, triggerType, summary]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [OPEN, SUPPRESSED, ESCALATION_FAILED]
        patientId: { type: string }
        deviceId:
          type: string
          nullable: true
        severity: { type: string }
        triggerType: { type: string }
        summary: { type: string }
      example: { id: "alt-701", status: "OPEN", patientId: "pat-1001", deviceId: "dev-22", severity: "HIGH", triggerType: "HEART_RATE", summary: "Heart rate exceeded configured threshold." }
...
    AppointmentEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/AppointmentResponse' }
    AppointmentListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/AppointmentResponse' }
    TeleconsultationEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/TeleconsultationResponse' }
    AvailableSlotEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/AvailableSlotResponse' }
...
    AlertEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/AlertResponse' }
    AlertListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/AlertResponse' }
...
    ApiError:
      type: object
      required: [correlationId, code, message, timestamp]
      properties:
        correlationId: { type: string }
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
      example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Patient record not found: pat-404", timestamp: "2026-05-31T10:10:00Z" }
  responses:
    BadRequestResponse:
      description: Validation or request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "VALIDATION_ERROR", message: "Validation failed", timestamp: "2026-05-31T10:11:00Z" }
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Resource not found", timestamp: "2026-05-31T10:11:20Z" }
    ConflictResponse:
      description: Business rule conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "SLOT_ALREADY_BOOKED", message: "Slot is already booked", timestamp: "2026-05-31T10:11:30Z" }
    InternalServerErrorResponse:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "INTERNAL_ERROR", message: "Unexpected server error", timestamp: "2026-05-31T10:11:40Z" }
security:
  - bearerAuth: []
paths:
...
  /appointments/available-slots:
    get:
      summary: List available slots for a provider on a specific date
      parameters:
        - in: query
          name: providerId
          required: true
          schema: { type: string }
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Available slots retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailableSlotEnvelope' }
              examples:
                availableSlots:
                  value:
                    correlationId: corr-123
                    data:
                      providerId: prov-44
                      date: 2026-06-01
                      availableSlots:
                        - 2026-06-01T09:00:00Z
                        - 2026-06-01T09:30:00Z
                        - 2026-06-01T10:00:00Z
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '503': { $ref: '#/components/responses/InternalServerErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /appointments:
    post:
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
            examples:
              bookAppointment:
                value:
                  patientId: pat-1001
                  providerId: prov-44
                  scheduledAt: 2026-06-01T09:30:00Z
                  channel: VIDEO
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppointmentEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: apt-2001
                      status: BOOKED
                      patientId: pat-1001
                      providerId: prov-44
                      scheduledAt: 2026-06-01T09:30:00Z
                      channel: VIDEO
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '503': { $ref: '#/components/responses/InternalServerErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List appointments
      responses:
        '200':
          description: Appointment collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppointmentListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: apt-2001
                        status: BOOKED
                        patientId: pat-1001
                        providerId: prov-44
                        scheduledAt: 2026-06-01T09:30:00Z
                        channel: VIDEO
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /appointments/{id}:
    get:
      summary: Retrieve an appointment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Appointment retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppointmentEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: apt-2001
                      status: BOOKED
                      patientId: pat-1001
                      providerId: prov-44
                      scheduledAt: 2026-06-01T09:30:00Z
                      channel: VIDEO
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /appointments/{id}/teleconsult/start:
    post:
      summary: Start teleconsultation session for appointment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Teleconsultation session started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeleconsultationEnvelope' }
              examples:
                started:
                  value:
                    correlationId: corr-123
                    data:
                      sessionId: tcs-5001
                      appointmentId: apt-2001
                      status: INITIATED
                      doctorJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=DOCTOR
                      patientJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=PATIENT
                      startedAt: 2026-06-20T10:25:00Z
                      joinedAt:
                      completedAt:
                      consultationNotes:
                      followUpRequired: false
                      nextFollowUpDate:
                      interactionLogs:
                        - 2026-06-20T10:25:00Z | Teleconsultation initiated by doctor
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '409': { $ref: '#/components/responses/ConflictResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /appointments/{id}/teleconsult/join:
    post:
      summary: Join teleconsultation session as patient
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Teleconsultation session joined
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeleconsultationEnvelope' }
              examples:
                joined:
                  value:
                    correlationId: corr-123
                    data:
                      sessionId: tcs-5001
                      appointmentId: apt-2001
                      status: IN_PROGRESS
                      doctorJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=DOCTOR
                      patientJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=PATIENT
                      startedAt: 2026-06-20T10:25:00Z
                      joinedAt: 2026-06-20T10:30:30Z
                      completedAt:
                      consultationNotes:
                      followUpRequired: false
                      nextFollowUpDate:
                      interactionLogs:
                        - 2026-06-20T10:25:00Z | Teleconsultation initiated by doctor
                        - 2026-06-20T10:30:30Z | Patient joined teleconsultation session
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '409': { $ref: '#/components/responses/ConflictResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /appointments/{id}/teleconsult/complete:
    post:
      summary: Complete teleconsultation and update EHR notes
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeleconsultationCompleteRequest'
            examples:
              complete:
                value:
                  consultationNotes: Consultation completed and medication plan reviewed.
                  followUpRequired: true
                  nextFollowUpDate: 2026-07-15T11:00:00Z
      responses:
        '200':
          description: Teleconsultation completed and notes recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeleconsultationEnvelope' }
              examples:
                completed:
                  value:
                    correlationId: corr-123
                    data:
                      sessionId: tcs-5001
                      appointmentId: apt-2001
                      status: COMPLETED
                      doctorJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=DOCTOR
                      patientJoinUrl: https://teleconsult.healthcare.local/session/tcs-5001?role=PATIENT
                      startedAt: 2026-06-20T10:25:00Z
                      joinedAt: 2026-06-20T10:30:30Z
                      completedAt: 2026-06-20T11:05:10Z
                      consultationNotes: Consultation completed and medication plan reviewed.
                      followUpRequired: true
                      nextFollowUpDate: 2026-07-15T11:00:00Z
                      interactionLogs:
                        - 2026-06-20T10:25:00Z | Teleconsultation initiated by doctor
                        - 2026-06-20T10:30:30Z | Patient joined teleconsultation session
                        - 2026-06-20T11:05:10Z | Doctor completed teleconsultation and notes were recorded
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '409': { $ref: '#/components/responses/ConflictResponse' }
        '503': { $ref: '#/components/responses/InternalServerErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...
  /alerts:
    post:
      summary: Trigger a clinical alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRequest'
            examples:
              triggerAlert:
                value:
                  patientId: pat-1001
                  deviceId: dev-22
                  severity: HIGH
                  triggerType: VITAL_THRESHOLD
                  summary: Heart rate exceeded configured threshold.
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: alt-701
                      status: OPEN
                      patientId: pat-1001
                      deviceId: dev-22
                      severity: HIGH
                      triggerType: VITAL_THRESHOLD
                      summary: Heart rate exceeded configured threshold.
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List alerts
      responses:
        '200':
          description: Alert collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: alt-701
                        status: OPEN
                        patientId: pat-1001
                        deviceId: dev-22
                        severity: HIGH
                        triggerType: VITAL_THRESHOLD
                        summary: Heart rate exceeded configured threshold.
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /alerts/{id}:
    get:
      summary: Retrieve an alert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Alert retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: alt-701
                      status: OPEN
                      patientId: pat-1001
                      deviceId: dev-22
                      severity: HIGH
                      triggerType: VITAL_THRESHOLD
                      summary: Heart rate exceeded configured threshold.
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...